requirements.txt
pydicom==2.1.2
Pillow==8.2.0
numpy==1.19.5